#1789 Restructured some bits and pieces to create a new RouteCreateService
