+++ app/Service/CombatLog/CreateRouteDungeonRouteServiceInterface.php
@@ -7,22 +7,8 @@ use App\Service\CombatLog\Models\CreateRoute\CreateRouteBody;
 use App\Service\CombatLog\ResultEvents\BaseResultEvent;
 use Illuminate\Support\Collection;
 
-interface CombatLogDungeonRouteServiceInterface
+interface CreateRouteDungeonRouteServiceInterface
 {
-    /**
-     * @param string $combatLogFilePath
-     *
-     * @return Collection|BaseResultEvent[]
-     */
-    public function getResultEvents(string $combatLogFilePath): Collection;
-
-    /**
-     * @param string $combatLogFilePath
-     *
-     * @return Collection|DungeonRoute[]
-     */
-    public function convertCombatLogToDungeonRoutes(string $combatLogFilePath): Collection;
-
     /**
      * @param CreateRouteBody $createRouteBody
      * @return DungeonRoute
