+++ app/Service/CombatLog/ResultEventDungeonRouteServiceInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Service\CombatLog;
+
+use App\Models\DungeonRoute;
+use Illuminate\Support\Collection;
+
+interface ResultEventDungeonRouteServiceInterface
+{
+    /**
+     * @param string $combatLogFilePath
+     *
+     * @return Collection|DungeonRoute[]
+     */
+    public function convertCombatLogToDungeonRoutes(string $combatLogFilePath): Collection;
+}
