+++ database/seeders/dungeondata/bfa/freehold/dungeonroutes.json
@@ -99,7 +99,8 @@
                     654,
                     655,
                     656
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -110,7 +111,8 @@
                 "index": 2,
                 "enemies": [
                     4223
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -124,7 +126,8 @@
                     677,
                     678,
                     679
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -135,7 +138,8 @@
                 "index": 4,
                 "enemies": [
                     176
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -149,7 +153,8 @@
                     986,
                     987,
                     988
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -162,7 +167,8 @@
                     681,
                     682,
                     683
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -173,7 +179,8 @@
                 "index": 7,
                 "enemies": [
                     4169
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -193,7 +200,8 @@
                     716,
                     717,
                     1030
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -213,7 +221,8 @@
                     1039,
                     1040,
                     1041
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -226,7 +235,8 @@
                     177,
                     180,
                     1036
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -243,7 +253,8 @@
                     2230,
                     2231,
                     3399
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -259,7 +270,8 @@
                     1110,
                     3402,
                     3403
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -274,7 +286,8 @@
                     1102,
                     3393,
                     3394
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -285,7 +298,8 @@
                 "index": 14,
                 "enemies": [
                     179
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -304,7 +318,8 @@
                     1099,
                     1100,
                     2228
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -315,7 +330,8 @@
                 "index": 16,
                 "enemies": [
                     3347
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -326,7 +342,8 @@
                 "index": 17,
                 "enemies": [
                     4224
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -337,7 +354,8 @@
                 "index": 18,
                 "enemies": [
                     178
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -353,7 +371,8 @@
                     744,
                     745,
                     746
-                ]
+                ],
+                "spells": []
             }
         ],
         "enemyraidmarkers": [],
