+++ database/seeders/dungeondata/bfa/shrineofthestorm/dungeonroutes.json
@@ -130,7 +130,8 @@
                     1123,
                     1124,
                     3121
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -156,7 +157,8 @@
                     2096,
                     3122,
                     3123
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -177,7 +179,8 @@
                     2100,
                     2101,
                     2102
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -188,7 +191,8 @@
                 "index": 4,
                 "enemies": [
                     4227
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -214,7 +218,8 @@
                     2136,
                     2137,
                     2138
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -238,7 +243,8 @@
                     2124,
                     2125,
                     2126
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -249,7 +255,8 @@
                 "index": 7,
                 "enemies": [
                     169
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -264,7 +271,8 @@
                     1148,
                     1149,
                     1150
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -280,7 +288,8 @@
                     1154,
                     1155,
                     1156
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -297,7 +306,8 @@
                     1161,
                     1162,
                     1163
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -312,7 +322,8 @@
                     1166,
                     1167,
                     1168
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -326,7 +337,8 @@
                     1170,
                     1171,
                     1172
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -337,7 +349,8 @@
                 "index": 13,
                 "enemies": [
                     4173
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -352,7 +365,8 @@
                     1176,
                     1179,
                     1180
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -372,7 +386,8 @@
                     1188,
                     1189,
                     1190
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -390,7 +405,8 @@
                     1196,
                     1197,
                     1198
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -408,7 +424,8 @@
                     1204,
                     1205,
                     1206
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -420,7 +437,8 @@
                 "enemies": [
                     168,
                     1207
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -433,7 +451,8 @@
                     1177,
                     1178,
                     1403
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -444,7 +463,8 @@
                 "index": 20,
                 "enemies": [
                     4174
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -458,7 +478,8 @@
                     1218,
                     2163,
                     2166
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -469,7 +490,8 @@
                 "index": 22,
                 "enemies": [
                     4228
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -480,7 +502,8 @@
                 "index": 23,
                 "enemies": [
                     170
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -506,7 +529,8 @@
                     1286,
                     1288,
                     1290
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -552,7 +576,8 @@
                     3135,
                     3137,
                     3138
-                ]
+                ],
+                "spells": []
             }
         ],
         "enemyraidmarkers": [],
