+++ database/seeders/dungeondata/bfa/siegeofboralus/dungeonroutes.json
@@ -114,7 +114,8 @@
                     3450,
                     3451,
                     3454
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -132,7 +133,8 @@
                     3478,
                     3479,
                     3480
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -150,7 +152,8 @@
                     3469,
                     3470,
                     3471
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -174,7 +177,8 @@
                     1984,
                     2014,
                     2015
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -185,7 +189,8 @@
                 "index": 5,
                 "enemies": [
                     4229
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -203,7 +208,8 @@
                     1988,
                     1989,
                     1991
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -229,7 +235,8 @@
                     1996,
                     1997,
                     1998
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -252,7 +259,8 @@
                     2002,
                     2003,
                     2004
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -265,7 +273,8 @@
                     167,
                     3475,
                     3476
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -281,7 +290,8 @@
                     1866,
                     2010,
                     2011
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -299,7 +309,8 @@
                     1873,
                     1874,
                     1875
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -310,7 +321,8 @@
                 "index": 12,
                 "enemies": [
                     4177
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -325,7 +337,8 @@
                     1878,
                     1879,
                     1880
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -339,7 +352,8 @@
                     1882,
                     1883,
                     1884
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -369,7 +383,8 @@
                     2036,
                     2037,
                     2038
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -389,7 +404,8 @@
                     2181,
                     2182,
                     2184
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -406,7 +422,8 @@
                     1895,
                     2045,
                     2046
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -417,7 +434,8 @@
                 "index": 18,
                 "enemies": [
                     4178
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -432,7 +450,8 @@
                     1903,
                     1904,
                     3424
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -447,7 +466,8 @@
                     1900,
                     1901,
                     1902
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -458,7 +478,8 @@
                 "index": 21,
                 "enemies": [
                     166
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -472,7 +493,8 @@
                     1906,
                     1907,
                     1908
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -483,7 +505,8 @@
                 "index": 23,
                 "enemies": [
                     4230
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -498,7 +521,8 @@
                     1932,
                     1933,
                     1934
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -512,7 +536,8 @@
                     1936,
                     1937,
                     1938
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -527,7 +552,8 @@
                     1941,
                     1942,
                     1943
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -540,7 +566,8 @@
                     1944,
                     1945,
                     1946
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -551,7 +578,8 @@
                 "index": 28,
                 "enemies": [
                     165
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -562,7 +590,8 @@
                 "index": 29,
                 "enemies": [
                     164
-                ]
+                ],
+                "spells": []
             }
         ],
         "enemyraidmarkers": [],
