+++ database/seeders/dungeondata/legion/lowerkarazhan/dungeonroutes.json
@@ -52,7 +52,8 @@
                     16645,
                     16646,
                     16647
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -69,7 +70,8 @@
                     16652,
                     16653,
                     16654
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -88,7 +90,8 @@
                     16671,
                     16672,
                     16673
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -100,7 +103,8 @@
                 "enemies": [
                     16674,
                     16675
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -118,7 +122,8 @@
                     16681,
                     16682,
                     16683
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -137,7 +142,8 @@
                     16690,
                     16691,
                     16692
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -175,7 +181,8 @@
                     16881,
                     16882,
                     17006
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -186,7 +193,8 @@
                 "index": 8,
                 "enemies": [
                     16883
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -205,7 +213,8 @@
                     16852,
                     16853,
                     16855
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -228,7 +237,8 @@
                     16835,
                     16836,
                     16854
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -243,7 +253,8 @@
                     16924,
                     16926,
                     16927
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -260,7 +271,8 @@
                     16911,
                     16912,
                     16913
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -275,7 +287,8 @@
                     16921,
                     16922,
                     16925
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -289,7 +302,8 @@
                     16916,
                     16917,
                     16918
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -304,7 +318,8 @@
                     16904,
                     16905,
                     16906
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -315,7 +330,8 @@
                 "index": 16,
                 "enemies": [
                     17007
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -332,7 +348,8 @@
                     16761,
                     16762,
                     16820
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -353,7 +370,8 @@
                     16770,
                     16771,
                     16772
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -373,7 +391,8 @@
                     16742,
                     16743,
                     16744
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -388,7 +407,8 @@
                     16733,
                     16734,
                     16750
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -401,7 +421,8 @@
                     16786,
                     16787,
                     16791
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -416,7 +437,8 @@
                     16794,
                     16803,
                     16804
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -430,7 +452,8 @@
                     16806,
                     16807,
                     16808
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -441,7 +464,8 @@
                 "index": 24,
                 "enemies": [
                     16819
-                ]
+                ],
+                "spells": []
             }
         ],
         "enemyraidmarkers": [],
