+++ database/seeders/dungeondata/shadowlands/hallsofatonement_a/dungeonroutes.json
@@ -104,7 +104,8 @@
                     4409,
                     4410,
                     4411
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -121,7 +122,8 @@
                     4372,
                     4373,
                     4453
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -134,7 +136,8 @@
                     4456,
                     4457,
                     4458
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -148,7 +151,8 @@
                     4376,
                     4377,
                     4378
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -164,7 +168,8 @@
                     4358,
                     4359,
                     4360
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -176,7 +181,8 @@
                 "enemies": [
                     4355,
                     4356
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -194,7 +200,8 @@
                     4439,
                     4440,
                     4455
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -208,7 +215,8 @@
                     4442,
                     4443,
                     4444
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -221,7 +229,8 @@
                     4425,
                     4426,
                     4427
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -236,7 +245,8 @@
                     4405,
                     4406,
                     4407
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -247,7 +257,8 @@
                 "index": 11,
                 "enemies": [
                     4399
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -261,7 +272,8 @@
                     4386,
                     4428,
                     4429
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -278,7 +290,8 @@
                     4463,
                     4814,
                     4815
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -289,7 +302,8 @@
                 "index": 14,
                 "enemies": [
                     4817
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -300,7 +314,8 @@
                 "index": 15,
                 "enemies": [
                     4818
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -315,7 +330,8 @@
                     4821,
                     4822,
                     4823
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -329,7 +345,8 @@
                     4756,
                     4757,
                     4758
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -340,7 +357,8 @@
                 "index": 18,
                 "enemies": [
                     4761
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -375,7 +393,8 @@
                     5453,
                     5454,
                     5455
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -386,7 +405,8 @@
                 "index": 20,
                 "enemies": [
                     4830
-                ]
+                ],
+                "spells": []
             }
         ],
         "enemyraidmarkers": [],
