+++ database/seeders/dungeondata/shadowlands/mistsoftirnescithe/dungeonroutes.json
@@ -65,7 +65,8 @@
                     4560,
                     4561,
                     4562
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -79,7 +80,8 @@
                     4564,
                     4565,
                     4566
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -93,7 +95,8 @@
                     4577,
                     4578,
                     4579
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -105,7 +108,8 @@
                 "enemies": [
                     4580,
                     4581
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -121,7 +125,8 @@
                     5093,
                     5094,
                     5095
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -135,7 +140,8 @@
                     5164,
                     5165,
                     5166
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -149,7 +155,8 @@
                     5097,
                     5098,
                     5099
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -163,7 +170,8 @@
                     5072,
                     5073,
                     5074
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -178,7 +186,8 @@
                     5169,
                     5170,
                     5188
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -193,7 +202,8 @@
                     5820,
                     5821,
                     5822
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -207,7 +217,8 @@
                     5100,
                     5101,
                     5102
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -225,7 +236,8 @@
                     5118,
                     5119,
                     5120
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -241,7 +253,8 @@
                     5124,
                     5125,
                     5126
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -258,7 +271,8 @@
                     5132,
                     5134,
                     5797
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -272,7 +286,8 @@
                     5136,
                     5137,
                     5138
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -290,7 +305,8 @@
                     5066,
                     5067,
                     5068
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -301,7 +317,8 @@
                 "index": 17,
                 "enemies": [
                     4545
-                ]
+                ],
+                "spells": []
             }
         ],
         "enemyraidmarkers": [],
