+++ database/seeders/dungeondata/shadowlands/necroticwake_a/dungeonroutes.json
@@ -95,7 +95,8 @@
                     4620,
                     4621,
                     4622
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -112,7 +113,8 @@
                     4631,
                     4632,
                     4633
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -127,7 +129,8 @@
                     4636,
                     4637,
                     4638
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -143,7 +146,8 @@
                     4642,
                     4643,
                     4646
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -160,7 +164,8 @@
                     4663,
                     4705,
                     4706
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -172,7 +177,8 @@
                 "enemies": [
                     4686,
                     4687
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -188,7 +194,8 @@
                     4695,
                     4696,
                     4697
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -202,7 +209,8 @@
                     4699,
                     4700,
                     4701
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -214,7 +222,8 @@
                 "enemies": [
                     4702,
                     4703
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -234,7 +243,8 @@
                     4679,
                     4680,
                     4681
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -247,7 +257,8 @@
                     4671,
                     4672,
                     4673
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -259,7 +270,8 @@
                 "enemies": [
                     4682,
                     4683
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -276,7 +288,8 @@
                     4711,
                     4712,
                     4713
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -287,7 +300,8 @@
                 "index": 14,
                 "enemies": [
                     4684
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -309,7 +323,8 @@
                     4732,
                     4733,
                     4740
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -324,7 +339,8 @@
                     4716,
                     4717,
                     4718
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -337,7 +353,8 @@
                     4719,
                     4720,
                     4721
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -349,7 +366,8 @@
                 "enemies": [
                     4738,
                     4739
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -363,7 +381,8 @@
                     4735,
                     4736,
                     4737
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -375,7 +394,8 @@
                 "enemies": [
                     4742,
                     4743
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -389,7 +409,8 @@
                     4745,
                     4746,
                     4747
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -403,7 +424,8 @@
                     4749,
                     4750,
                     4751
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -415,7 +437,8 @@
                 "enemies": [
                     4741,
                     5842
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -426,7 +449,8 @@
                 "index": 24,
                 "enemies": [
                     4752
-                ]
+                ],
+                "spells": []
             }
         ],
         "enemyraidmarkers": [],
