+++ database/seeders/dungeondata/shadowlands/plaguefall/dungeonroutes.json
@@ -64,7 +64,8 @@
                     4863,
                     4864,
                     4867
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -77,7 +78,8 @@
                     4910,
                     4911,
                     4912
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -91,7 +93,8 @@
                     4917,
                     4918,
                     4919
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -122,7 +125,8 @@
                     4993,
                     4994,
                     4996
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -133,7 +137,8 @@
                 "index": 5,
                 "enemies": [
                     4965
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -148,7 +153,8 @@
                     4833,
                     5002,
                     5003
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -162,7 +168,8 @@
                     5001,
                     5011,
                     5012
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -176,7 +183,8 @@
                     5005,
                     5006,
                     5882
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -190,7 +198,8 @@
                     5008,
                     5009,
                     5010
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -201,7 +210,8 @@
                 "index": 10,
                 "enemies": [
                     5024
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -216,7 +226,8 @@
                     5021,
                     5022,
                     5023
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -228,7 +239,8 @@
                 "enemies": [
                     5027,
                     5029
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -240,7 +252,8 @@
                 "enemies": [
                     5028,
                     5030
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -251,7 +264,8 @@
                 "index": 14,
                 "enemies": [
                     5031
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -264,7 +278,8 @@
                     4871,
                     4872,
                     4873
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -277,7 +292,8 @@
                     4868,
                     4869,
                     4870
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -289,7 +305,8 @@
                 "enemies": [
                     4875,
                     4876
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -300,7 +317,8 @@
                 "index": 18,
                 "enemies": [
                     4874
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -312,7 +330,8 @@
                 "enemies": [
                     4878,
                     5034
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -328,7 +347,8 @@
                     5039,
                     5042,
                     5043
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -341,7 +361,8 @@
                     4883,
                     4886,
                     4887
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -355,7 +376,8 @@
                     4889,
                     4890,
                     4933
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -369,7 +391,8 @@
                     4892,
                     4893,
                     4894
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -380,7 +403,8 @@
                 "index": 24,
                 "enemies": [
                     4840
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -398,7 +422,8 @@
                     4949,
                     4951,
                     4952
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -413,7 +438,8 @@
                     4956,
                     4957,
                     4958
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -428,7 +454,8 @@
                     4962,
                     4963,
                     4964
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -439,7 +466,8 @@
                 "index": 29,
                 "enemies": [
                     4839
-                ]
+                ],
+                "spells": []
             }
         ],
         "enemyraidmarkers": [],
