+++ database/seeders/dungeondata/shadowlands/sanguinedepths_a/dungeonroutes.json
@@ -96,7 +96,8 @@
                     5379,
                     5380,
                     5381
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -109,7 +110,8 @@
                     5382,
                     5383,
                     5384
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -121,7 +123,8 @@
                 "enemies": [
                     5389,
                     5865
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -136,7 +139,8 @@
                     5402,
                     5843,
                     5866
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -154,7 +158,8 @@
                     5394,
                     5395,
                     5396
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -165,7 +170,8 @@
                 "index": 6,
                 "enemies": [
                     5373
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -182,7 +188,8 @@
                     5468,
                     5469,
                     5470
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -193,7 +200,8 @@
                 "index": 8,
                 "enemies": [
                     5471
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -206,7 +214,8 @@
                     5479,
                     5480,
                     5481
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -219,7 +228,8 @@
                     5482,
                     5483,
                     5484
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -236,7 +246,8 @@
                     5490,
                     5491,
                     5492
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -249,7 +260,8 @@
                     5511,
                     5512,
                     5513
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -265,7 +277,8 @@
                     5517,
                     5518,
                     5519
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -278,7 +291,8 @@
                     5520,
                     5521,
                     5522
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -292,7 +306,8 @@
                     5528,
                     5529,
                     5530
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -305,7 +320,8 @@
                     5565,
                     5566,
                     5567
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -319,7 +335,8 @@
                     5534,
                     5535,
                     5536
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -332,7 +349,8 @@
                     5539,
                     5540,
                     5541
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -343,7 +361,8 @@
                 "index": 19,
                 "enemies": [
                     5545
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -356,7 +375,8 @@
                     5574,
                     5575,
                     5576
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -368,7 +388,8 @@
                 "enemies": [
                     5572,
                     5573
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -379,7 +400,8 @@
                 "index": 22,
                 "enemies": [
                     5371
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -395,7 +417,8 @@
                     5580,
                     5581,
                     5582
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -408,7 +431,8 @@
                     5583,
                     5584,
                     5585
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -423,7 +447,8 @@
                     5869,
                     5870,
                     5871
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -439,7 +464,8 @@
                     5596,
                     5597,
                     5598
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -454,7 +480,8 @@
                     5874,
                     5875,
                     5876
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -469,7 +496,8 @@
                     5610,
                     5611,
                     5612
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -483,7 +511,8 @@
                     5879,
                     5880,
                     5881
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -496,7 +525,8 @@
                     5616,
                     5617,
                     5618
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -514,7 +544,8 @@
                     5601,
                     5619,
                     5620
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -525,7 +556,8 @@
                 "index": 32,
                 "enemies": [
                     5372
-                ]
+                ],
+                "spells": []
             }
         ],
         "enemyraidmarkers": [],
