+++ database/seeders/dungeondata/shadowlands/tazaveshsoleahsgambit/dungeonroutes.json
@@ -59,7 +59,8 @@
                     7240,
                     7241,
                     7242
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -76,7 +77,8 @@
                     7222,
                     7223,
                     7224
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -94,7 +96,8 @@
                     7328,
                     7329,
                     7481
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -112,7 +115,8 @@
                     7335,
                     7336,
                     7337
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -131,7 +135,8 @@
                     7344,
                     7345,
                     7482
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -150,7 +155,8 @@
                     7352,
                     7353,
                     7486
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -169,7 +175,8 @@
                     7402,
                     7403,
                     7404
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -188,7 +195,8 @@
                     7360,
                     7361,
                     7362
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -200,7 +208,8 @@
                 "enemies": [
                     7405,
                     7406
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -212,7 +221,8 @@
                 "enemies": [
                     7407,
                     7408
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -223,7 +233,8 @@
                 "index": 11,
                 "enemies": [
                     7409
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -240,7 +251,8 @@
                     7417,
                     7418,
                     7419
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -254,7 +266,8 @@
                     7422,
                     7423,
                     7424
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -268,7 +281,8 @@
                     7426,
                     7427,
                     7428
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -282,7 +296,8 @@
                     7431,
                     7432,
                     7433
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -293,7 +308,8 @@
                 "index": 16,
                 "enemies": [
                     7429
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -306,7 +322,8 @@
                     7435,
                     7436,
                     7437
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -321,7 +338,8 @@
                     7443,
                     7444,
                     7445
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -334,7 +352,8 @@
                     7438,
                     7439,
                     7440
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -345,7 +364,8 @@
                 "index": 20,
                 "enemies": [
                     7446
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -356,7 +376,8 @@
                 "index": 21,
                 "enemies": [
                     7447
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -367,7 +388,8 @@
                 "index": 22,
                 "enemies": [
                     7448
-                ]
+                ],
+                "spells": []
             }
         ],
         "enemyraidmarkers": [],
