+++ database/seeders/dungeondata/shadowlands/theaterofpain/dungeonroutes.json
@@ -98,7 +98,8 @@
                     5175,
                     5176,
                     5177
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -111,7 +112,8 @@
                     5179,
                     5180,
                     5181
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -122,7 +124,8 @@
                 "index": 3,
                 "enemies": [
                     5182
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -133,7 +136,8 @@
                 "index": 4,
                 "enemies": [
                     5861
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -148,7 +152,8 @@
                     5185,
                     5186,
                     5187
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -159,7 +164,8 @@
                 "index": 6,
                 "enemies": [
                     5257
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -170,7 +176,8 @@
                 "index": 7,
                 "enemies": [
                     5178
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -186,7 +193,8 @@
                     5224,
                     5225,
                     5227
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -202,7 +210,8 @@
                     5230,
                     5231,
                     5859
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -213,7 +222,8 @@
                 "index": 10,
                 "enemies": [
                     5232
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -227,7 +237,8 @@
                     5234,
                     5235,
                     5236
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -240,7 +251,8 @@
                     5237,
                     5238,
                     5239
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -252,7 +264,8 @@
                 "enemies": [
                     5246,
                     5247
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -265,7 +278,8 @@
                     5621,
                     5622,
                     5623
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -276,7 +290,8 @@
                 "index": 15,
                 "enemies": [
                     5256
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -292,7 +307,8 @@
                     5193,
                     5194,
                     5195
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -308,7 +324,8 @@
                     5199,
                     5200,
                     5201
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -321,7 +338,8 @@
                     5202,
                     5203,
                     5204
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -332,7 +350,8 @@
                 "index": 19,
                 "enemies": [
                     5206
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -347,7 +366,8 @@
                     5209,
                     5210,
                     5211
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -358,7 +378,8 @@
                 "index": 21,
                 "enemies": [
                     5212
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -374,7 +395,8 @@
                     5216,
                     5217,
                     5218
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -386,7 +408,8 @@
                 "enemies": [
                     5219,
                     5220
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -397,7 +420,8 @@
                 "index": 24,
                 "enemies": [
                     5189
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -408,7 +432,8 @@
                 "index": 25,
                 "enemies": [
                     5174
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -419,7 +444,8 @@
                 "index": 26,
                 "enemies": [
                     5857
-                ]
+                ],
+                "spells": []
             }
         ],
         "enemyraidmarkers": [],
