+++ database/seeders/dungeondata/wod/irondocks/dungeonroutes.json
@@ -48,7 +48,8 @@
                     7884,
                     7885,
                     7896
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -62,7 +63,8 @@
                     7887,
                     7888,
                     7889
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -76,7 +78,8 @@
                     7894,
                     7895,
                     7916
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -90,7 +93,8 @@
                     7913,
                     7914,
                     7915
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -104,7 +108,8 @@
                     7911,
                     7917,
                     7918
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -118,7 +123,8 @@
                     7920,
                     7921,
                     7922
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -130,7 +136,8 @@
                 "enemies": [
                     7923,
                     8525
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -141,7 +148,8 @@
                 "index": 8,
                 "enemies": [
                     8493
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -162,7 +170,8 @@
                     7958,
                     7959,
                     7960
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -184,7 +193,8 @@
                     7956,
                     7957,
                     7980
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -206,7 +216,8 @@
                     7965,
                     7966,
                     7967
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -219,7 +230,8 @@
                     7968,
                     7969,
                     7970
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -233,7 +245,8 @@
                     7972,
                     7973,
                     7974
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -248,7 +261,8 @@
                     7977,
                     7978,
                     7979
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -259,7 +273,8 @@
                 "index": 15,
                 "enemies": [
                     7980
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -273,7 +288,8 @@
                     7982,
                     7983,
                     7984
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -286,7 +302,8 @@
                     7985,
                     7986,
                     7987
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -298,7 +315,8 @@
                 "enemies": [
                     7988,
                     7989
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -312,7 +330,8 @@
                     7991,
                     7992,
                     7993
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -324,7 +343,8 @@
                 "enemies": [
                     7997,
                     7998
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -338,7 +358,8 @@
                     8000,
                     8001,
                     8002
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -351,7 +372,8 @@
                     8007,
                     8008,
                     8009
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -362,7 +384,8 @@
                 "index": 23,
                 "enemies": [
                     8010
-                ]
+                ],
+                "spells": []
             },
             {
                 "floor_id": null,
@@ -375,7 +398,8 @@
                     8011,
                     8012,
                     8013
-                ]
+                ],
+                "spells": []
             }
         ],
         "enemyraidmarkers": [],
